videogen: yt-dlp clip downloader + README polish
videogen/download_clips.py
  Thin yt-dlp wrapper that turns a ChronoMagic-Pro parquet manifest into
  a directory of mp4s ingest_chronomagic --video-dir can consume.
  Restart-safe (already-downloaded ids are skipped), filters out clips
  longer than --max-duration, supports a naive caption-substring filter
  for grabbing just a phase-transition slice without downloading the
  whole 465K-clip corpus.  Errors per id are tolerated; the rest of the
  pipeline treats missing files as "no video bytes for this row".

dataloader.py
  Force a writable buffer in the fp16 → torch conversion so PyTorch
  doesn't warn "given NumPy array is not writable" under multi-worker
  loading.  Confirmed multi-worker safe (350 batches/s at num_workers=2
  on the smoke table).

README.md
  Pulled the measured numbers (B5: 7.58× speedup, +25.3 MFU pts;
  B10: 240 s on 8 clips) into a "Headline numbers" block at the top so
  they're visible without grepping commit messages.  Cross-linked the
  bench scripts.  Added an Evaluation section showing how to invoke the
  ChronoMagic + VBench scaffolds against a saved checkpoint.

+++ examples/videogen/videogen/dataloader.py
@@ -111,7 +111,8 @@ def _flat_fp16_to_tensor(arr: pa.Array, dtype: torch.dtype, shape: tuple[int, ..
     np_dtype = np_dtype_map[dtype]
     # zero_copy_only=False because the FSL → flat numpy conversion does one
     # contiguous copy.  Still 10-100× cheaper than per-row decode + VAE.
-    flat = arr.flatten().to_numpy(zero_copy_only=False).astype(np_dtype, copy=False)
+    # ``copy=True`` forces a writable buffer so torch.from_numpy doesn't warn.
+    flat = arr.flatten().to_numpy(zero_copy_only=False).astype(np_dtype, copy=True)
     flat = flat.reshape((len(arr),) + shape)
     return torch.from_numpy(flat)
